Xcos: discard style update on interfacefunction change
On update, style is already updated so there is no real need apart from
manual/debug testing within Xcos. This style update introduce style
reset while (on Scilab Thread) setting an interface function on a block
(trigger a notification passed to the EDT) *and* setting the style. On
the next EDT execution, the style will be discarded !

Change-Id: Ia0fbc352b5137e32ae4ab5eae9d2a3a4b9be5c62